refactor: move uniformize in utils
